refactor: move ToolsRepositoryHelper back to tool
 so cyclic imports are prevented
