+++ cyclonedx/model/vulnerability.py
@@ -41,7 +41,7 @@ from sortedcontainers import SortedSet
 from .._internal.compare import ComparableTuple as _ComparableTuple
 from ..exception.model import MutuallyExclusivePropertiesException, NoPropertiesProvidedException
 from ..schema.schema import SchemaVersion1Dot4, SchemaVersion1Dot5, SchemaVersion1Dot6
-from ..serialization import BomRefHelper, ToolsRepositoryHelper
+from ..serialization import BomRefHelper
 from . import Property, XsUri
 from .bom_ref import BomRef
 from .contact import OrganizationalContact, OrganizationalEntity
@@ -51,7 +51,7 @@ from .impact_analysis import (
     ImpactAnalysisResponse,
     ImpactAnalysisState,
 )
-from .tool import Tool, ToolsRepository
+from .tool import Tool, ToolsRepository, _ToolsRepositoryHelper
 
 
 @serializable.serializable_class
@@ -1247,7 +1247,7 @@ class Vulnerability:
         self._credits = credits
 
     @property
-    @serializable.type_mapping(ToolsRepositoryHelper)
+    @serializable.type_mapping(_ToolsRepositoryHelper)
     @serializable.xml_sequence(17)
     def tools(self) -> ToolsRepository:
         """
